backend/requirements-django.txt
django==5.2.13
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-allauth==65.16.0
django-cors-headers==4.9.0
dj-database-url==3.1.2
pydantic-settings==2.13.1
